backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Website cloning service using LLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "playwright>=1.42.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.3",
    "uvicorn>=0.27.1",
    "python-multipart>=0.0.9",
    "beautifulsoup4>=4.12.3",
    "requests>=2.31.0",
]
